+++ qcluster/describer.py
@@ -1,12 +1,15 @@
 import os
 
-import ollama
+from ollama import Client
 
 from qcluster.custom_types import ClusterDescription
 
 
 def get_description(document: str):
-  response = ollama.chat(
+  model = os.environ['OLLAMA_MODEL']
+  client = Client(host=os.environ['OLLAMA_HOST'])
+
+  response = client.chat(
     messages=[
       {
         'role': 'user',
@@ -15,8 +18,9 @@ def get_description(document: str):
           f' and a short title for the following document:\n\n{document}',
       }
     ],
-    model=os.environ['OLLAMA_MODEL'],
+    model=model,
     format=ClusterDescription.model_json_schema(),
+    options={'temperature': 0.0},
   )
 
   return ClusterDescription.model_validate_json(response.message.content)